#4092 - Move getTimeInCurrentTimezone to util
