+++ packages/scandipwa/src/component/MyAccountOrder/MyAccountOrder.component.js
@@ -18,6 +18,7 @@ import MyAccountOrderInformation from 'Component/MyAccountOrderInformation';
 import MyAccountOrderItemsTable from 'Component/MyAccountOrderItemsTable';
 import MyAccountOrderTabs from 'Component/MyAccountOrderTabs';
 import { OrderType } from 'Type/Order.type';
+import { getTimeInCurrentTimezone } from 'Util/Form/Extract';
 import { convertStringToDate } from 'Util/Manipulations/Date';
 
 import {
@@ -39,8 +40,7 @@ export class MyAccountOrder extends PureComponent {
         rss_order_subscribe_allow: PropTypes.bool.isRequired,
         handleChangeActiveTab: PropTypes.func.isRequired,
         activeTab: PropTypes.string.isRequired,
-        isMobile: PropTypes.bool.isRequired,
-        getTimeInCurrentTimezone: PropTypes.func.isRequired
+        isMobile: PropTypes.bool.isRequired
     };
 
     renderMap = {
@@ -181,7 +181,7 @@ export class MyAccountOrder extends PureComponent {
     }
 
     renderOrderComments() {
-        const { activeTab, order: { comments = [] }, getTimeInCurrentTimezone } = this.props;
+        const { activeTab, order: { comments = [] } } = this.props;
 
         if (activeTab !== ORDER_ITEMS || !comments.length) {
             return null;
